Prep rename to lowercase lexer
